Major Refactoring:
Move parser&decoder into a single unit. Make the parser a private helper
function of the decoder. As the decoders and parsers are now one unit per
log, they will be referred to as parsers.

Create a master Parse that returns a writable stream that also emits the
events for that writable stream. This Parse delegates to the correct
parser for the particular log file.